add JspMealController, roles test

+++ src/main/java/ru/javawebinar/topjava/repository/jdbc/JdbcUserRepository.java
@@ -2,18 +2,25 @@ package ru.javawebinar.topjava.repository.jdbc;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.support.DataAccessUtils;
+import org.springframework.jdbc.core.BatchPreparedStatementSetter;
 import org.springframework.jdbc.core.BeanPropertyRowMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
+import org.springframework.jdbc.support.rowset.SqlRowSet;
 import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
+import ru.javawebinar.topjava.model.Role;
 import ru.javawebinar.topjava.model.User;
 import ru.javawebinar.topjava.repository.UserRepository;
 
-import java.util.List;
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+import java.util.*;
 
 @Repository
+@Transactional(readOnly = true)
 public class JdbcUserRepository implements UserRepository {
 
     private static final BeanPropertyRowMapper<User> ROW_MAPPER = BeanPropertyRowMapper.newInstance(User.class);
@@ -35,12 +42,27 @@ public class JdbcUserRepository implements UserRepository {
     }
 
     @Override
+    @Transactional
     public User save(User user) {
         BeanPropertySqlParameterSource parameterSource = new BeanPropertySqlParameterSource(user);
 
         if (user.isNew()) {
             Number newKey = insertUser.executeAndReturnKey(parameterSource);
-            user.setId(newKey.intValue());
+            int newId = newKey.intValue();
+            user.setId(newId);
+            List<Role> roles = new ArrayList<>(user.getRoles());
+            jdbcTemplate.batchUpdate("INSERT INTO user_roles VALUES (?, ?)", new BatchPreparedStatementSetter() {
+                @Override
+                public void setValues(PreparedStatement ps, int i) throws SQLException {
+                    ps.setInt(1, newId);
+                    ps.setString(2, roles.get(i).name());
+                }
+
+                @Override
+                public int getBatchSize() {
+                    return roles.size();
+                }
+            });
         } else if (namedParameterJdbcTemplate.update(
                 "UPDATE users SET name=:name, email=:email, password=:password, " +
                         "registered=:registered, enabled=:enabled, calories_per_day=:caloriesPerDay WHERE id=:id", parameterSource) == 0) {
@@ -50,13 +72,15 @@ public class JdbcUserRepository implements UserRepository {
     }
 
     @Override
+    @Transactional
     public boolean delete(int id) {
-        return jdbcTemplate.update("DELETE FROM users WHERE id=?", id) != 0;
+        return jdbcTemplate.update("DELETE FROM user_roles WHERE user_id=?", id) != 0 && jdbcTemplate.update("DELETE FROM users WHERE id=?", id) != 0;
     }
 
     @Override
     public User get(int id) {
         List<User> users = jdbcTemplate.query("SELECT * FROM users WHERE id=?", ROW_MAPPER, id);
+        users.forEach(this::setRolesToUser);
         return DataAccessUtils.singleResult(users);
     }
 
@@ -64,11 +88,33 @@ public class JdbcUserRepository implements UserRepository {
     public User getByEmail(String email) {
 //        return jdbcTemplate.queryForObject("SELECT * FROM users WHERE email=?", ROW_MAPPER, email);
         List<User> users = jdbcTemplate.query("SELECT * FROM users WHERE email=?", ROW_MAPPER, email);
+        users.forEach(this::setRolesToUser);
         return DataAccessUtils.singleResult(users);
     }
 
     @Override
     public List<User> getAll() {
-        return jdbcTemplate.query("SELECT * FROM users ORDER BY name, email", ROW_MAPPER);
+        List<User> users = jdbcTemplate.query("SELECT * FROM users ORDER BY name, email", ROW_MAPPER);
+        SqlRowSet sqlRowSetRoles = jdbcTemplate.queryForRowSet("SELECT * FROM user_roles");
+        Map<Integer, Set<Role>> roles = mapOfRoles(sqlRowSetRoles);
+        users.forEach(user -> user.setRoles(roles.getOrDefault(user.getId(), Collections.emptySet())));
+        return users;
+    }
+
+    private void setRolesToUser(User user) {
+        SqlRowSet sqlRowSet = jdbcTemplate.queryForRowSet("SELECT * FROM user_roles WHERE user_id=?", user.getId());
+        Map<Integer, Set<Role>> roles = mapOfRoles(sqlRowSet);
+        user.setRoles(roles.getOrDefault(user.getId(), Collections.emptySet()));
+    }
+
+    private Map<Integer, Set<Role>> mapOfRoles(SqlRowSet sqlRowSet) {
+        Map<Integer, Set<Role>> rolesMap = new HashMap<>();
+        while (sqlRowSet.next()) {
+            int id = sqlRowSet.getInt("user_id");
+            Role role = Role.valueOf(sqlRowSet.getString("role"));
+            rolesMap.putIfAbsent(id, new HashSet<>());
+            rolesMap.get(id).add(role);
+        }
+        return rolesMap;
     }
 }
